Move `h.views.api_exceptions` -> `h.views.api.exceptions`
